+++ runtime/context.go
@@ -30,11 +30,16 @@ import (
 
 type contextFieldKey string
 
+// ContextScopeTagsExtractor defines func where extracts tags from context
+type ContextScopeTagsExtractor func(context.Context) map[string]string
+
 const (
-	endpointKey        = contextFieldKey("endpoint")
-	requestUUIDKey     = contextFieldKey("requestUUID")
-	routingDelegateKey = contextFieldKey("rd")
-	requestLogFields   = contextFieldKey("requestLogFields")
+	endpointKey           = contextFieldKey("endpoint")
+	requestUUIDKey        = contextFieldKey("requestUUID")
+	routingDelegateKey    = contextFieldKey("rd")
+	endpointRequestHeader = contextFieldKey("endpointRequestHeader")
+	requestLogFields      = contextFieldKey("requestLogFields")
+	requestScopeFields    = contextFieldKey("requestScopeFields")
 )
 
 const (
@@ -47,6 +52,20 @@ const (
 	logFieldRequestUUID         = "requestUUID"
 	logFieldEndpointID          = "endpointID"
 	logFieldHandlerID           = "handlerID"
+)
+
+const (
+	scopeFieldRequestMethod = "method"
+	scopeFieldEndpointID    = "endpointID"
+	scopeFieldHandlerID     = "handlerID"
+)
+
+const (
+	// EndpointScope defines the name of endpoint scope
+	EndpointScope = "endpoint"
+
+	// ClientScope defines the name of client scope
+	ClientScope = "client"
 )
 
 // WithEndpointField adds the endpoint information in the
@@ -62,6 +81,30 @@ func GetRequestEndpointFromCtx(ctx context.Context) string {
 		return endpoint
 	}
 	return ""
+}
+
+// WithEndpointRequestHeadersField adds the endpoint request header information in the
+// request context.
+func WithEndpointRequestHeadersField(ctx context.Context, requestHeaders map[string]string) context.Context {
+	headers := GetEndpointRequestHeadersFromCtx(ctx)
+	for k, v := range requestHeaders {
+		headers[k] = v
+	}
+
+	return context.WithValue(ctx, endpointRequestHeader, headers)
+}
+
+// GetEndpointRequestHeadersFromCtx returns the endpoint request headers, if it exists on context
+func GetEndpointRequestHeadersFromCtx(ctx context.Context) map[string]string {
+	requestHeaders := make(map[string]string)
+	if val := ctx.Value(endpointRequestHeader); val != nil {
+		headers, _ := val.(map[string]string)
+		for k, v := range headers {
+			requestHeaders[k] = v
+		}
+	}
+
+	return requestHeaders
 }
 
 // withRequestFields annotates zanzibar request context to context.Context. In
@@ -105,6 +148,29 @@ func WithLogFields(ctx context.Context, newFields ...zap.Field) context.Context 
 	return context.WithValue(ctx, requestLogFields, accumulateLogFields(ctx, newFields))
 }
 
+// WithScopeFields returns a new context with the given scope fields attached to context.Context
+func WithScopeFields(ctx context.Context, newFields map[string]string) context.Context {
+	fields := GetScopeFieldsFromCtx(ctx)
+	for k, v := range newFields {
+		fields[k] = v
+	}
+
+	return context.WithValue(ctx, requestScopeFields, fields)
+}
+
+// GetScopeFieldsFromCtx returns the tag info extracted from context.
+func GetScopeFieldsFromCtx(ctx context.Context) map[string]string {
+	fields := make(map[string]string)
+	if val := ctx.Value(requestScopeFields); val != nil {
+		headers, _ := val.(map[string]string)
+		for k, v := range headers {
+			fields[k] = v
+		}
+	}
+
+	return fields
+}
+
 func accumulateLogFields(ctx context.Context, newFields []zap.Field) []zap.Field {
 	previousFieldsUntyped := ctx.Value(requestLogFields)
 	if previousFieldsUntyped == nil {
@@ -122,6 +188,41 @@ func accumulateLogFields(ctx context.Context, newFields []zap.Field) []zap.Field
 	}
 
 	return fields
+}
+
+// ContextExtractor is a extractor that extracts some log fields from the context
+type ContextExtractor interface {
+	ExtractScopeTags(ctx context.Context) map[string]string
+}
+
+// AddContextScopeTagsExtractor added a scope tags extractor to contextExtractor.
+func (c *ContextExtractors) AddContextScopeTagsExtractor(extractors ...ContextScopeTagsExtractor) {
+	c.contextScopeExtractors = extractors
+}
+
+// MakeContextExtractor returns a extractor that extracts log fields a context.
+func (c *ContextExtractors) MakeContextExtractor() ContextExtractor {
+	return &ContextExtractors{
+		contextScopeExtractors: c.contextScopeExtractors,
+	}
+}
+
+// ContextExtractors warps extractors for context
+type ContextExtractors struct {
+	contextScopeExtractors []ContextScopeTagsExtractor
+}
+
+// ExtractScopeTags extracts scope fields from a context into a tag.
+func (c *ContextExtractors) ExtractScopeTags(ctx context.Context) map[string]string {
+	tags := make(map[string]string)
+	for _, fn := range c.contextScopeExtractors {
+		sc := fn(ctx)
+		for k, v := range sc {
+			tags[k] = v
+		}
+	}
+
+	return tags
 }
 
 // ContextLogger is a logger that extracts some log fields from the context before passing through to underlying zap logger.